requirements.txt
django==2.1.2
django-crispy-forms==1.7.2
coverage==4.5.1